+++ evaluations/evaluate_csv.py
@@ -1,10 +1,11 @@
 import functools
 
+from grammar_graph import gg
 from grammar_graph.gg import GrammarGraph
 
 from isla.fuzzer import GrammarFuzzer
 from isla.performance_evaluator import Evaluator
-from isla.solver import ISLaSolver, CostSettings, CostWeightVector
+from isla.solver import ISLaSolver, CostSettings, CostWeightVector, GrammarBasedBlackboxCostComputer
 from isla_formalizations import csv
 
 max_number_free_instantiations = 10
@@ -14,7 +15,6 @@ eval_k = 3
 # cost_vector = STD_COST_SETTINGS.weight_vectors[0]
 cost_vector = CostWeightVector(
     tree_closing_cost=1,
-    vacuous_penalty=0,
     constraint_cost=0,
     derivation_depth_penalty=1,
     low_k_coverage_penalty=0,
@@ -26,7 +26,9 @@ g_colno = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k),
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(csv.CSV_GRAMMAR)),
     global_fuzzer=False,
     fuzzer_factory=functools.partial(GrammarFuzzer, min_nonterminals=0, max_nonterminals=30),
 )