+++ evaluations/evaluate_xml.py
@@ -1,7 +1,8 @@
+from grammar_graph import gg
 from grammar_graph.gg import GrammarGraph
 
 from isla.performance_evaluator import Evaluator
-from isla.solver import ISLaSolver, CostSettings, CostWeightVector
+from isla.solver import ISLaSolver, CostSettings, CostWeightVector, GrammarBasedBlackboxCostComputer
 from isla_formalizations import xml_lang
 
 max_number_free_instantiations = 10
@@ -10,7 +11,6 @@ eval_k = 4
 
 cost_vector = CostWeightVector(
     tree_closing_cost=15,
-    vacuous_penalty=0,
     constraint_cost=0,
     derivation_depth_penalty=0,
     low_k_coverage_penalty=5,
@@ -22,7 +22,9 @@ g_wf = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k)
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(xml_lang.XML_GRAMMAR_WITH_NAMESPACE_PREFIXES)),
 )
 
 g_ns = lambda timeout: ISLaSolver(
@@ -31,7 +33,9 @@ g_ns = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k)
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(xml_lang.XML_GRAMMAR_WITH_NAMESPACE_PREFIXES)),
 )
 
 g_redef = lambda timeout: ISLaSolver(
@@ -40,7 +44,9 @@ g_redef = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k)
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(xml_lang.XML_GRAMMAR_WITH_NAMESPACE_PREFIXES)),
 )
 
 g_wf_ns = lambda timeout: ISLaSolver(
@@ -49,7 +55,9 @@ g_wf_ns = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k)
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(xml_lang.XML_GRAMMAR_WITH_NAMESPACE_PREFIXES)),
 )
 
 g_wf_ns_redef = lambda timeout: ISLaSolver(
@@ -60,7 +68,9 @@ g_wf_ns_redef = lambda timeout: ISLaSolver(
     max_number_free_instantiations=max_number_free_instantiations,
     max_number_smt_instantiations=max_number_smt_instantiations,
     timeout_seconds=timeout,
-    cost_settings=CostSettings(cost_vector, k=eval_k)
+    cost_computer=GrammarBasedBlackboxCostComputer(
+        CostSettings(cost_vector, k=eval_k),
+        gg.GrammarGraph.from_grammar(xml_lang.XML_GRAMMAR_WITH_NAMESPACE_PREFIXES)),
 )
 
 if __name__ == '__main__':